refactor: view 와 viewModel 의존성 분리
