+++ lib/screen/sign_up/reset_pw_screen.dart
@@ -4,7 +4,9 @@ import 'package:provider/provider.dart';
 import 'package:rest_api_ex/config/common/sized_box_values.dart';
 import 'package:rest_api_ex/config/constants.dart';
 import 'package:rest_api_ex/config/custom_app_bar.dart';
+import 'package:rest_api_ex/config/custom_snack_bar.dart';
 import 'package:rest_api_ex/config/custom_submit_button.dart';
+import 'package:rest_api_ex/config/navigate_to.dart';
 import 'package:rest_api_ex/config/user_info_text_form_field.dart';
 import 'package:rest_api_ex/config/validation_check.dart';
 import 'package:rest_api_ex/design/color_styles.dart';
@@ -12,39 +14,50 @@ import 'package:rest_api_ex/design/font_styles.dart';
 import 'package:rest_api_ex/screen/my_bottom_navigation.dart';
 import 'package:rest_api_ex/screen/view_model/reset_pw_view_model.dart';
 
-class ResetPwScreen extends StatelessWidget {
+class ResetPwScreen extends StatefulWidget {
   const ResetPwScreen({super.key});
 
   @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (_) => ResetPwViewModel(),
-      builder: (context, child) {
-        final viewModel = context.watch<ResetPwViewModel>();
+  State<ResetPwScreen> createState() => _ResetPwScreenState();
+}
 
-        return Scaffold(
-          appBar: const CustomAppBar(title: Constants.resetPw),
-          body: ModalProgressHUD(
-            inAsyncCall: viewModel.showSpinner,
-            child: Form(
-              key: viewModel.formKey,
-              child: Padding(
-                padding: const EdgeInsets.all(20.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    _description(),
-                    SizedBoxValues.gapH20,
-                    _inputPassword(viewModel),
-                    SizedBoxValues.gapH20,
-                    _changePassword(viewModel, context),
-                  ],
-                ),
-              ),
+class _ResetPwScreenState extends State<ResetPwScreen> {
+  final _userPasswordController = TextEditingController();
+  final _userPasswordConfirmController = TextEditingController();
+
+  @override
+  void dispose() {
+    _userPasswordController.dispose();
+    _userPasswordConfirmController.dispose();
+
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = context.watch<ResetPwViewModel>();
+
+    return Scaffold(
+      appBar: const CustomAppBar(title: Constants.resetPw),
+      body: ModalProgressHUD(
+        inAsyncCall: viewModel.showSpinner,
+        child: Form(
+          key: viewModel.formKey,
+          child: Padding(
+            padding: const EdgeInsets.all(20.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                _description(),
+                SizedBoxValues.gapH20,
+                _inputPassword(viewModel),
+                SizedBoxValues.gapH20,
+                _changePassword(viewModel, context),
+              ],
             ),
           ),
-        );
-      },
+        ),
+      ),
     );
   }
 
@@ -76,7 +89,7 @@ class ResetPwScreen extends StatelessWidget {
 
         // 비밀번호
         UserInfoTextFormField(
-          controller: viewModel.userPasswordController,
+          controller: _userPasswordController,
           validator: validatePassword,
           decorationLabelText: Constants.pwValidationRule,
         ),
@@ -98,10 +111,10 @@ class ResetPwScreen extends StatelessWidget {
 
         // 비밀번호 확인
         UserInfoTextFormField(
-          controller: viewModel.userPasswordConfirmController,
+          controller: _userPasswordConfirmController,
           validator: (value) => validateConfirmPassword(
             value,
-            viewModel.userPasswordController.text,
+            _userPasswordController.text,
           ),
           decorationLabelText: Constants.pwValidationRule,
         ),
@@ -114,7 +127,11 @@ class ResetPwScreen extends StatelessWidget {
       formKey: viewModel.formKey,
       buttonText: Constants.submitNewPw,
       nextPage: const MyBottomNavigation(),
-      onPressedCallback: () => viewModel.submitNewPassword(context, 'password'),
+      // onPressedCallback: () => viewModel.submitNewPassword(context, 'password'),
+      onPressedCallback: () {
+        navigatePushAndRemoveUtilTo(context, const MyBottomNavigation());
+        CustomSnackBar().showSnackBar(context, '비밀번호가 변경되었습니다.');
+      },
     );
   }
 }
